chore(packer.lua): enable auto_attach in nvim-navbuddy setup feat(telescope-file-browser.lua): load the telescope file_browser extension refactor(telescope.lua): remove unnecessary telescope extensions The `auto_attach` option in the `lsp` configuration of `nvim-navbuddy` is now set to `true`, enabling automatic attachment of the LSP client. This improves the functionality of `nvim-navbuddy` by automatically attaching to the LSP server.
The `telescope-file-browser` extension is now loaded in the `telescope-file-browser.lua` file. This allows the use of the `file_browser` extension in Telescope, providing a file browser functionality within Telescope.

In the `telescope.lua` file, unnecessary Telescope extensions have been removed. The `coc`, `gh`, `project`, and `file_browser` extensions are no longer loaded. This simplifies the configuration and reduces potential conflicts or unnecessary overhead.